Fix infinite rendering calling questions bug
