+++ frontend/components/Question.tsx
@@ -30,8 +30,10 @@ const Question = ({
   setIsCreate,
   colorMode,
   userMode,
+  questions,
+  fetchQuestions,
 }) => {
-  const [questions, setQuestions] = useState(null);
+  // const [questions, setQuestions] = useState(null);
   const [description, setDescription] = useState(null);
   const [difficulty, setDifficulty] = useState(null);
   const [category, setCategory] = useState(null);
@@ -40,19 +42,19 @@ const Question = ({
   const [openQuestion, setOpenQuestion] = useState(null);
   const { isOpen, onOpen, onClose } = useDisclosure();
 
-  const fetchQuestions = async () => {
-    try {
-      const res = await axios.get(`${IP_ADDRESS}:3001/questions/getall`);
+  // const fetchQuestions = async () => {
+  //   try {
+  //     const res = await axios.get(`${IP_ADDRESS}:3001/questions/getall`);
 
-      setQuestions(res.data.qns);
-    } catch (error) {
-      console.log("ERROR: ", error);
-    }
-  };
+  //     setQuestions(res.data.qns);
+  //   } catch (error) {
+  //     console.log("ERROR: ", error);
+  //   }
+  // };
 
-  useEffect(() => {
-    fetchQuestions();
-  }, [questions]);
+  // useEffect(() => {
+  //   fetchQuestions();
+  // }, []);
 
   const handleEdit = async ({
     qn_num,
