+++ frontend/components/QuestionsInputField.tsx
@@ -28,8 +28,8 @@ const QuestionInputField = ({
   isCreate,
   setIsCreate,
   colorMode,
+  setQuestions,
 }) => {
-  const [question, setQuestions] = useState(null);
   const [error, setError] = useState(false);
 
   const [selectedComplexity, setSelectedComplexity] = useState("");
@@ -88,7 +88,6 @@ const QuestionInputField = ({
     });
 
     setIsCreate(true);
-
     fetchQuestions();
   };
